+++ VR_Polytechnique/Assets/Prefabs/Menu.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7db4e8c1787a4284bbe9155ad051019b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
