+++ VR_Polytechnique/Assets/Prefabs/Menu/QuitButton.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c010d618f229105408a53c5a6843239d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
