+++ VR_Polytechnique/Assets/Prefabs/Menu/StartButton.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f3bfe33266ef2d14587fd7a3cea97b1b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
