+++ VR_Polytechnique/Assets/Prefabs/Menu/TutorielButton.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0ff25aec3a824784c96b5975829346fd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
