+++ VR_Polytechnique/Assets/Prefabs/Menu/VolumeSlider.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b450e93f89fb3384ba7e25af0bde560a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
