--- VR_Polytechnique/Assets/Scripts/StartGameButton.cs
@@ -1,14 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-public class StartGameButton : MonoBehaviour
-{
-
-    public int gameStartScene;
-
-    public void StartGame()
-    {
-        SceneManager.LoadScene(gameStartScene);
-    }
-}
