--- VR_Polytechnique/Assets/Scripts/StartGameButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6b1ff2daa221508458f697778dfbc5df
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
