--- VR_Polytechnique/Assets/Scripts/TextSlider.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 18fca0432d30533469aac7b687aaa21b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
