+++ VR_Polytechnique/UserSettings/EditorUserSettings.asset
@@ -21,10 +21,10 @@ EditorUserSettings:
       value: 55550652060d5f0e0e5f087546720b44174f1c727a7f24657b2b4431b3e43660
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 5b01515503025a59585d5f76467609444e16482b747e27627b2b4f66b1e6303b
+      value: 54540150550558080b5f5d2145755d44174e497e297975317c714965e4e3673d
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 54540150550558080b5f5d2145755d44174e497e297975317c714965e4e3673d
+      value: 5b01515503025a59585d5f76467609444e16482b747e27627b2b4f66b1e6303b
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
